+++ master.py
@@ -11,8 +11,7 @@ from modules.groundtruth_generator_module import generator
 from modules.segmentation_module import segmentation
 from modules.cell_tracking_module import cell_tracking
 from modules.graph_filtering_module import graph_filtering
-from modules.graph_event_filter_module import graph_event_filter
-from modules.fusion_correction_module import fusion_correction
+from modules.events_filter_module import events_filter
 from modules.viewer_module import viewer
 from general import general_functions as gf
 import multiprocessing as mp
@@ -85,14 +84,7 @@ class GraphEventFilter(gf.Page):
     def __init__(self):
         super().__init__()
         self.window = QVBoxLayout(self.container)
-        self.window.addWidget(graph_event_filter.GraphEventFilter())
-        self.window.addStretch()
-
-class FusionCorrection(gf.Page):
-    def __init__(self):
-        super().__init__()
-        self.window = QVBoxLayout(self.container)
-        self.window.addWidget(fusion_correction.FusionCorrection())
+        self.window.addWidget(events_filter.GraphEventFilter())
         self.window.addStretch()
 
 
@@ -129,8 +121,7 @@ class MainWindow(QWidget):
         tabwizard.addPage(Segmentation(), "Segmentation")
         tabwizard.addPage(CellTracking(), "Cell tracking")
         tabwizard.addPage(GraphFiltering(), "Graph filtering")
-        tabwizard.addPage(GraphEventFilter(), "Graph event filter")
-        tabwizard.addPage(FusionCorrection(), "Fusion correction")
+        tabwizard.addPage(GraphEventFilter(), "Events filter")
 
         layout = QHBoxLayout()
         self.status_line = QLineEdit()